Add quick access to dice macros from dice tool
