+++ client/src/game/systems/customData/components/DiceFormat.vue
@@ -2,23 +2,28 @@
 import { computed } from "vue";
 
 import type { ApiShapeCustomDataDiceExpression } from "../../../../apiTypes";
+import type { LocalId } from "../../../../core/id";
 import { ToolName } from "../../../models/tools";
 import { activateTool } from "../../../tools/tools";
 import { diceToolInput } from "../../../tools/variants/dice";
 import { diceSystem } from "../../dice";
+import { diceState } from "../../dice/state";
+import { DiceUiState } from "../../dice/types";
 import type { ElementId } from "../types";
 import { getVariableSegments } from "../utils";
 
-const { element } = defineProps<{
+const { element, shapeFocus } = defineProps<{
     element: ApiShapeCustomDataDiceExpression & { id: ElementId };
+    shapeFocus?: LocalId;
 }>();
 
 const segments = computed(() => {
-    return getVariableSegments(element.value);
+    return getVariableSegments(element.value, shapeFocus);
 });
 
 async function loadRoll(): Promise<void> {
     if (!diceSystem.isLoaded) await diceSystem.loadSystems();
+    diceState.mutableReactive.uiState = DiceUiState.Roll;
     activateTool(ToolName.Dice);
     const system = diceSystem.getSystem("2d")!;
     // We need a short timeout when loading the dice tool for the first time
