+++ client/src/game/systems/dice/state.ts
@@ -1,10 +1,14 @@
 import { type Part, type RollResult, type SYSTEMS } from "@planarally/dice/core";
 import type { DeepReadonly } from "vue";
 
+import type { LocalId } from "../../../core/id";
 import { buildState } from "../../../core/systems/state";
 import type { AsyncReturnType } from "../../../core/types";
 
+import { DiceUiState } from "./types";
+
 interface DiceState {
+    uiState: LocalId | DiceUiState;
     dimensions3d: { width: number; height: number };
     history: { roll: RollResult<Part>; name: string; player: string }[];
     systems?: { "2d": AsyncReturnType<typeof SYSTEMS.DX>["DX"]; "3d": AsyncReturnType<typeof SYSTEMS.DX3>["DX3"] };
@@ -12,6 +16,7 @@ interface DiceState {
 }
 
 const state = buildState<DiceState>({
+    uiState: DiceUiState.Roll,
     dimensions3d: { width: 0, height: 0 },
     history: [],
 });
