+++ client/src/game/systems/dice/types.ts
@@ -1,0 +1,4 @@
+export enum DiceUiState {
+    Roll,
+    Macro,
+}
